Extract peer logic to new class
