+++ lib/server.js
@@ -5,20 +5,23 @@ const MessageFactory = require('./message-factory.js')
 const debug = require('debug')('slsk:server:i')
 
 let stack = require('./stack')
-let conn
 
 module.exports = class Server extends EventEmitter {
   constructor (cb) {
     super()
-    conn = net.createConnection({
+    this.conn = net.createConnection({
       host: 'server.slsknet.org',
       port: 2242
     }, cb)
 
     let msgs = new Messages()
 
-    conn.on('error', err => {
+    this.conn.on('error', err => {
       cb(new Error(err.message))
+    })
+
+    this.conn.on('data', data => {
+      msgs.write(data)
     })
 
     msgs.on('message', msg => {
@@ -96,26 +99,22 @@ module.exports = class Server extends EventEmitter {
         }
       }
     })
-
-    conn.on('data', data => {
-      msgs.write(data)
-    })
   }
 
   login (credentials) {
-    conn.write(MessageFactory
+    this.conn.write(MessageFactory
       .to.server
       .login(credentials).getBuff())
   }
 
   fileSearch (query, token) {
-    conn.write(MessageFactory
+    this.conn.write(MessageFactory
       .to.server
       .fileSearch(query, token)
       .getBuff())
   }
 
   destroy () {
-    conn.destroy()
+    this.conn.destroy()
   }
 }
