+++ lib/slsk-client.js
@@ -1,17 +1,13 @@
 const EventEmitter = require('events')
-
-const net = require('net')
 const crypto = require('crypto')
-const zlib = require('zlib')
 const stream = require('stream')
 
 const debug = require('debug')('slsk:i')
 const debugPeer = require('debug')('slsk:peer:i')
 
-const Messages = require('./messages.js')
-const MessageFactory = require('./message-factory.js')
 const downloadPeerFile = require('./download-peer-file.js')
 const Server = require('./server.js')
+const Peer = require('./peer.js')
 let stack = require('./stack')
 
 let server
@@ -35,119 +31,9 @@ class SlskClient extends EventEmitter {
       return
     }
 
-    let conn = net.createConnection({
-      host: peer.host,
-      port: peer.port
-    }, () => {
-      let buf = Buffer.from('05000000' + '00' + peer.token, 'hex')
-      conn.write(buf)
-      peers[peer.user] = conn
-    })
-
-    conn.on('error', error => {
-      debug(`${peer.user} error ${error.code}`)
-    })
-
-    let msgs = new Messages()
-
-    msgs.on('message', msg => {
-      let size = msg.int32()
-      debugPeer(`${peer.user} size: ${size}`)
-      if (size <= 4) return
-
-      let code = msg.int32()
-      switch (code) {
-        case 4: {
-          debugPeer(`recv GetSharedFileList`)
-          let res = MessageFactory
-            .to.peer
-            .sharedFileList(1, 'test', 1, '@@abcde\\test\\test-test.mp3')
-          debugPeer(`sending SharedFileList`)
-          conn.write(res.getBuff())
-          break
-        }
-        case 9: {
-          debugPeer(`recv FileSearchResult`)
-          // This command use zlib for communication, we must decompress it before using it
-          let content = msg.data.slice(msg.pointer, size + 4)
-          zlib.unzip(content, (err, buffer) => {
-            if (err) {
-              debugPeer(err)
-              return
-            }
-
-            let result = MessageFactory.from.peer.fileSearchResult(buffer)
-
-            if (stack.search[result.currentToken]) {
-              result.files.forEach(file => {
-                stack.search[result.currentToken].cb({
-                  user: file.user,
-                  file: file.file,
-                  size: file.size,
-                  slots: result.slots === 1,
-                  bitrate: file.attribs ? file.attribs[0] : undefined,
-                  speed: result.speed
-                })
-              })
-            }
-          })
-          break
-        }
-        case 40: {
-          let dir = msg.int32()
-          let token = msg.rawHexStr(4)
-          let file = msg.str()
-          debugPeer(`recv TransferRequest ${file}`)
-          stack.downloadTokens[token] = {
-            user: peer.user,
-            file
-          }
-          if (dir === 1) {
-            stack.downloadTokens[token].size = msg.int32()
-          }
-          let res = MessageFactory
-            .to.peer
-            .transferResponse(token)
-
-          setTimeout(() => {
-            debugPeer(`sending TransferResponse`)
-            conn.write(res.getBuff())
-          }, 200)
-          break
-        }
-        case 41: {
-          let token = msg.rawHexStr(4)
-          let allowed = msg.int8()
-          debugPeer(`recv TransferResponse token: ${token} allowed: ${allowed}`)
-          if (allowed === 0) {
-            let reason = msg.str()
-            debugPeer(`reason: ${reason}, I will receive TransferRequest soon...`)
-            delete stack.downloadTokens[token] // avoid memory leak
-          } else if (allowed === 1) {
-            debugPeer(`Directly allowed. Connecting to peer with PeerInit + ${token}`)
-            downloadPeerFile(peer.host, peer.port, peer.token, peer.user, true)
-          }
-          break
-        }
-        case 46: {
-          let filename = msg.str()
-          debugPeer(`UploadFailed ${filename}`)
-          let down = stack.download[peer.user + '_' + filename]
-          if (down && typeof down.cb === 'function') {
-            down.cb(new Error('Peer error'))
-          } else {
-            debugPeer(`Cannot cb for ${peer.user} ${filename}`)
-          }
-          break
-        }
-        default: {
-          debugPeer(`unknown peer message code ${code}`)
-        }
-      }
-    })
-
-    conn.on('data', data => {
-      msgs.write(data)
+    let peerObj = new Peer(peer)
+    peerObj.on('connected', () => {
+      peers[peer.user] = peerObj
     })
   }
 
@@ -182,24 +68,18 @@ class SlskClient extends EventEmitter {
   }
 
   download (obj, cb, stream) {
-    debug(`launch download ${obj.user} ${obj.file}`)
+    debug(`launch download ${obj.file.user} ${obj.file.file}`)
     if (typeof cb !== 'function' && !stream) {
       throw new Error('2nd argument must be callback function')
     }
     if (typeof obj.file === 'undefined') {
       throw new Error('You must specify file')
     }
-
     if (!peers[obj.file.user]) {
-      // console.log(obj)
-      // console.log(Object.keys(peers))
       return cb(new Error('User not exist'))
     }
-    let token = crypto.randomBytes(4).toString('hex')
 
-    let msg = MessageFactory
-      .to.peer
-      .transferRequest(obj.file.file, token)
+    let token = crypto.randomBytes(4).toString('hex')
 
     stack.downloadTokens[token] = {
       user: obj.file.user,
@@ -212,7 +92,7 @@ class SlskClient extends EventEmitter {
       path: obj.path,
       stream
     }
-    peers[obj.file.user].write(msg.getBuff())
+    peers[obj.file.user].transferRequest(obj.file.file, token)
   }
 
   downloadStream (obj, cb) {
